requirements.txt
pygetwindow>=0.0.9
pyautogui>=0.9.54
pillow>=10.0.0
opencv-python-headless>=4.9.0.80
easyocr>=1.7.1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
